🎨 Retravail MessageService

+++ MessageMicroservice/MessageMicroservice.API/Controllers/MessagesController.cs
@@ -1,43 +1,57 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.Net.Mime;
 using System.Threading.Tasks;
 using MessageMicroservice.Domain.Models;
 using MessageMicroservice.Services.Services;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
-namespace MessageMicroservice.Controllers
+namespace MessageMicroservice.API.Controllers
 {
-    
     [Route("api/[controller]")]
     [ApiController]
+    [Produces(MediaTypeNames.Application.Json)]
+    [Consumes(MediaTypeNames.Application.Json)]
     public class MessagesController : ControllerBase
     {
-        private readonly MessageService _service;
+        private readonly IMessageService _service;
 
-        public MessagesController(MessageService service)
+        public MessagesController(IMessageService service)
         {
             _service = service;
         }
 
         // GET: api/<MessagesController>
         [HttpGet]
-        public async Task<List<Message>> GetAsync()
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<List<Message>> GetAll()
         {
-            return await _service.GetAsync();
+            return await _service.GetAllAsync();
         }
 
         // GET api/<MessagesController>/5
         [HttpGet("{id}")]
-        public async Task<Message> GetAsync(string id)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<Message>> GetById(string id)
         {
-            return await _service.GetByIdAsync(id);
+            var message = await _service.GetByIdAsync(id);
+            if (message == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(message);
         }
 
         // GET api/<MessagesController>/5
         [HttpGet("{guid:Guid}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<List<Message>> GetByGuidAsync(Guid guid)
         {
             return await _service.GetByGuidAsync(guid);
@@ -45,19 +59,33 @@ namespace MessageMicroservice.Controllers
 
         // POST api/<MessagesController>
         [HttpPost]
-        public async Task<IActionResult> Post([FromBody] Message message)
+        [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(Message))]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<ActionResult<Message>> Post(Message postMessage)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest("Invalid data");
             }
+
+            var message = new Message()
+            {
+                Guest = postMessage.Guest,
+                Author = postMessage.Author,
+                TimeStamp = postMessage.TimeStamp,
+                Body = postMessage.Body
+            };
+            
             await _service.Insert(message);
-            return Accepted();
+            return CreatedAtAction(nameof(GetById), new {id = message.Id}, message);
         }
 
         // PUT/PATCH api/<MessagesController>/5
         [HttpPut("{id}")]
         [HttpPatch("{id}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Put(string id, [FromBody] Message message)
         {
             var oldMessage = await _service.GetByIdAsync(id);
@@ -70,18 +98,21 @@ namespace MessageMicroservice.Controllers
             {
                 return BadRequest();
             }
-            var newMessage = await _service.UpdateByIdAsync(id, message);
-            return Accepted(newMessage);
+            var newMessage = await _service.UpdateById(id, message);
+            return NoContent();
         }
 
         // DELETE api/<MessagesController>
         [HttpDelete]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Delete()
         {      
             var result = await _service.DeleteAll();
             if (result)
             {
-                return Ok();
+                return NoContent();
             }
             else
             {
@@ -91,12 +122,20 @@ namespace MessageMicroservice.Controllers
 
         // DELETE api/<MessagesController>/5
         [HttpDelete("{id}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Delete(string id)
         {
+            var oldMessage = await _service.GetByIdAsync(id);
+            if (oldMessage == null)
+            {
+                return NotFound();
+            }
             var result = await _service.DeleteById(id);
             if (result)
             {
-                return Ok();
+                return NoContent();
             }
             else
             {
